perf: analytics + graph deep-dive benchmarks, DuckDB REAL->DOUBLE fidelity fix, performance report
- test_analytics: DuckDB-vs-SQLite per query shape (correctness + speedup),
  mirror sync overhead, full-sync cost
- test_graph: search_graph / search_graph_ppr (flagship), traverse depth 3,
  to_networkx build + cache hit, sync_graph_nodes, community_detect,
  betweenness; idempotent batch benchmarks
- scripts/analytics_sweep.py: DuckDB vs SQLite crossover sweep (10k/100k/1M)
- fix: DuckDB mirror stored REAL as float32 (DuckDB REAL), losing precision
  vs SQLite float64 (111.19 -> 111.19000244140625); now DOUBLE
- docs/PERFORMANCE.md: full study — accuracy (BEIR), analytics sweep, graph,
  search speed, recommendations
- key findings: mirror is 6-350x faster with no write overhead; filtered agg
  352x at 1M rows; point lookups 2.4x (no index in mirror); journal is the
  write bottleneck (~2k rows/s); search_graph/PPR dominates graph latency

+++ hybriddb/analytics.py
@@ -118,7 +118,7 @@ class AnalyticsMixin:
             sqlite_cols = [(row["name"], row["type"]) for row in cur.fetchall()]
         type_map = {
             "BOOLEAN": "INTEGER", "JSON": "TEXT", "TEXT": "TEXT",
-            "LONGTEXT": "TEXT", "INTEGER": "BIGINT", "REAL": "REAL",
+            "LONGTEXT": "TEXT", "INTEGER": "BIGINT", "REAL": "DOUBLE",
         }
         duck_cols: list[tuple[str, str]] = []
         for name, sqlite_type in sqlite_cols:
